Update Transit Assignment Procedure (#30)
* Added the ability to set what time transit assignments take place at.

* Switched to headway-transit with optimal search

+++ TMG.Visum/Utilities/MainZoneExtensions.cs
@@ -10,6 +10,6 @@ internal static class MainZoneExtensions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static int ZoneNumber(this IMainZone zone)
     {
-        return (int)zone.AttValue["No"];
+        return (int)(double)zone.AttValue["No"];
     }
 }
